Update 2.8.md
The call to E2.rvalue() possibly has the side effect of printing three-address code with temporaries, so you can't put it in emit().
You also forgot to emit the start label, which should be done before the call to E2.rvalue(), because the statement in the loop may have the side effect of changing the value of expression E2, so it possibly has to be re-evaluated (i.e. the three-address code with temporaries).